+++ server/component/style/textarea/js/textarea.js
@@ -1,3 +1,20 @@
+// jsonLogic export
+const jsonLogicOperators = {
+    field_equal: { op: '==' },
+    field_not_equal: { op: '!=' },
+    field_greater: { op: '>' },
+    field_less: { op: '<' },
+    equal: { op: '==' },
+    not_equal: { op: '!=' },
+    greater: { op: '>' },
+    less: { op: '<' },
+    less_or_equal: { op: '<=' },
+    greater_or_equal: { op: '>=' },
+    not_in: { op: '!=' },
+    in: { op: '==' },
+    in_one_of: { op: '==' },
+};
+
 dataConfigInitCalls = {};
 
 $(document).ready(function () {
@@ -7,34 +24,49 @@ $(document).ready(function () {
         // load the monaco editor for json fields
         require.config({ paths: { vs: BASE_PATH + '/js/ext/vs' } });
         var json = $(this)[0];
-        require(['vs/editor/editor.main'], function () {
-            var model = null;
-            if ($(json).prev().attr('name').includes('data_config')) {
-                model = setDataConfigSchema(monaco, json);
-            } else if ($(json).prev().attr('name').includes('condition')) {
-                model = setConditionSchema(monaco, json);
-            }
-            var editorOptions = {
-                value: $(json).prev().val(),
-                language: 'json',
-                automaticLayout: true,
-                renderLineHighlight: "none"
-            }
-            if (model) {
-                editorOptions['model'] = model;
-            }
-            var editor = monaco.editor.create(json, editorOptions);
-            editor.getAction('editor.action.formatDocument').run().then(() => {
-                calcMonacoEditorSize(editor, json);
+        if ($(json).prev().attr('name').includes('jquery_builder_json')) {
+            // this field is hidden and a holder only
+            $(json).parent().parent().addClass('d-none'); //hide the label
+            return;
+        } else {
+            require(['vs/editor/editor.main'], function () {
+                var model = null;
+                if ($(json).prev().attr('name').includes('data_config')) {
+                    model = setDataConfigSchema(monaco, json);
+                } else if ($(json).prev().attr('name').includes('condition')) {
+                    model = setConditionSchema(monaco, json);
+                }
+                var editorOptions = {
+                    value: $(json).prev().val(),
+                    language: 'json',
+                    automaticLayout: true,
+                    renderLineHighlight: "none"
+                }
+                if (model) {
+                    editorOptions['model'] = model;
+                }
+                var editor = monaco.editor.create(json, editorOptions);
+                editor.getAction('editor.action.formatDocument').run().then(() => {
+                    calcMonacoEditorSize(editor, json);
+                });
+                editor.onDidChangeModelContent(function (e) {
+                    $(json).prev().val(editor.getValue());
+                    calcMonacoEditorSize(editor, json);
+                });
+                if ($(json).prev().attr('name').includes('data_config')) {
+                    showDataConfiBuilder(json, editor);
+                }
+                if ($(json).prev().attr('name').includes('condition')) {
+                    var jquerBuilderJsonInput;
+                    $('textarea').each(function () {
+                        if ($(this).attr('name') && $(this).attr('name').includes('jquery_builder_json')) {
+                            jquerBuilderJsonInput = this;
+                        }
+                    })
+                    showConditionBuilder(json, editor, jquerBuilderJsonInput);
+                }
             });
-            editor.onDidChangeModelContent(function (e) {
-                $(json).prev().val(editor.getValue());
-                calcMonacoEditorSize(editor, json);
-            });
-            if ($(json).prev().attr('name').includes('data_config')) {
-                showDataConfiBuilder(json, editor);
-            }
-        });
+        }
     })
 });
 
@@ -74,7 +106,6 @@ function setDataConfigSchema(monaco, json) {
             }
         }]
     });
-
     return model;
 }
 
@@ -85,7 +116,7 @@ function setConditionSchema(monaco, json) {
     var model = monaco.editor.createModel($(json).prev().val(), "json", modelUri);
 
     // configure the JSON language support with schemas and schema associations
-    monaco.languages.json.jsonDefaults.setDiagnosticsOptions({
+    let r = monaco.languages.json.jsonDefaults.setDiagnosticsOptions({
         validate: true,
         enableSchemaRequest: true,
         schemas: [{
@@ -100,7 +131,6 @@ function setConditionSchema(monaco, json) {
             }
         }]
     });
-
     return model;
 }
 
@@ -119,9 +149,12 @@ function getJson(json) {
         return JSON.parse($(json).prev().val());
     } catch (error) {
         console.log('Error parsing the data_config value!');
-        return {};
+        return null;
     }
 }
+
+
+// ********************************************* DATA CONFIG BUILDER *****************************************
 
 // show the data config builder
 // on click the modal is loaded and show the builder
@@ -139,11 +172,14 @@ function showDataConfiBuilder(json, monacoEditor) {
                 editor.setValue(getJson(json));
             }
             $('.data_config_builder_modal_holder').on('hidden.bs.modal', function (e) {
-                // on modal close set the value to the Monaco editor
-                monacoEditor.getModel().setValue(JSON.stringify(editor.getValue(), null, 3));
+                
             })
-            $('.closeModal').each(function (){
+            $('.saveDataConfig').each(function () {
                 $(this).attr('data-dismiss', 'modal');
+                // on modal close set the value to the Monaco editor
+                $(this).click(function () {
+                    monacoEditor.getModel().setValue(JSON.stringify(editor.getValue(), null, 3));
+                })                
             });
         });
     });
@@ -173,7 +209,6 @@ function showDataConfiBuilder(json, monacoEditor) {
                             getTableNames(editor.getEditor(key).getValue(), editor, key.replace('type', 'table'));
                         } else if (key.includes('field_name')) {
                             // populate field names for the new field
-                            console.log('new field', key);
                             var keys = key.split('.');
                             var dataSourceKey = keys[0] + '.' + keys[1];
                             getTableFieldNames(
@@ -184,7 +219,7 @@ function showDataConfiBuilder(json, monacoEditor) {
                                 true
                             );
                         }
-                        editor.watch(key, watcherCallback.bind(editor, key)); // add the keys again - it is used to add watches for new sources
+                        editor.watch(key, dataConfigWatcherCallback.bind(editor, key)); // add the keys again - it is used to add watches for new sources
                     }
                 }
             })
@@ -216,7 +251,7 @@ function showDataConfiBuilder(json, monacoEditor) {
     });
 }
 
-const watcherCallback = function (path) {
+const dataConfigWatcherCallback = function (path) {
     if (path.includes('type') && this.getEditor(path) && this.getEditor(path).getValue()) {
         getTableNames(this.getEditor(path).getValue(), this, path.replace('type', 'table'));
     } else if (path.includes('table') && this.getEditor(path) && this.getEditor(path).getValue()) {
@@ -240,7 +275,7 @@ function checkAllDataConfigInitCalls(editor) {
         for (let key in editor.editors) {
             if (editor.editors.hasOwnProperty(key) && key !== 'root' &&
                 (key.includes('table') || key.includes('type') || key.includes('field_name'))) {
-                editor.watch(key, watcherCallback.bind(editor, key)); // add watch events to all exisitng keys containing [table, type, field_name]
+                editor.watch(key, dataConfigWatcherCallback.bind(editor, key)); // add watch events to all exisitng keys containing [table, type, field_name]
             }
         }
     }
@@ -327,4 +362,296 @@ function getTableFieldNames(type, formName, obj, path, init, editor) {
         },
         'json'
     );
+}
+// ********************************************* DATA CONFIG BUILDER *****************************************
+
+
+
+
+
+// ********************************************* CONDITION BUILDER *****************************************
+
+// AJAX call to get the table fields
+function getGroups(jquerBuilderJsonInput, monacoEditor) {
+    $.post(
+        BASE_PATH + '/request/AjaxDataSource/get_groups',
+        {},
+        function (data) {
+            if (data.success) {
+                var groups = [];
+                try {
+                    groups = JSON.parse(data.data);
+                } catch (error) {
+                    console.log('Error while parsing JSON', data.data);
+                }
+                prepareConditionBuilder(groups, jquerBuilderJsonInput, monacoEditor);
+            }
+            else {
+                console.log(data);
+            }
+        },
+        'json'
+    );
+}
+
+
+// prepare the condition builder and the rules that can be added
+function prepareConditionBuilder(groups, jquerBuilderJsonInput, monacoEditor) {
+
+    var queryStructure = {
+        icons: {
+            add_group: 'fas fa-plus-circle',
+            add_rule: 'fas fa-plus',
+            remove_group: 'fas fa-minus-square',
+            remove_rule: 'fas fa-minus-circle',
+            error: 'fas fa-exclamation-triangle',
+            sortable: 'fas fa-exclamation-triangle'
+        },
+        operators: $.fn.queryBuilder.constructor.DEFAULTS.operators.concat([
+            { type: 'field_equal', optgroup: '', nb_inputs: 2, multiple: false, apply_to: ['string'] },
+            { type: 'field_not_equal', optgroup: '', nb_inputs: 2, multiple: false, apply_to: ['string'] },
+            { type: 'field_greater', optgroup: '', nb_inputs: 2, multiple: false, apply_to: ['string'] },
+            { type: 'field_less', optgroup: '', nb_inputs: 2, multiple: false, apply_to: ['string'] },
+            { type: 'field_between', optgroup: '', nb_inputs: 3, multiple: false, apply_to: ['string'] },
+            { type: 'between_dates', optgroup: '', nb_inputs: 2, multiple: false, apply_to: ['date'] },
+            { type: 'in_one_of', optgroup: '', nb_inputs: 1, multiple: true, apply_to: ['array'] },
+        ]),
+        lang: {
+            operators: {
+                field_between: '[Field Name] between [value1(number)] and [value2(number)]',
+                field_greater: '[Field Name] > [value(number)]',
+                field_less: '[Field Name] < [value(number)]',
+                field_equal: '[Field Name] = [value]',
+                field_not_equal: '[Field Name] <> [value]',
+                between_dates: 'between',
+                in_one_of: 'in one of'
+            }
+        },
+        filters: [
+            {
+                id: 'user_group',
+                label: 'User group',
+                type: 'string',
+                input: 'select',
+                multiple: true,
+                values: groups,
+                plugin: 'selectpicker',
+                plugin_config: {
+                    liveSearch: true,
+                    width: 'auto',
+                    liveSearchStyle: 'contains',
+                },
+                operators: ['in', 'not_in', 'in_one_of']
+            }, {
+                id: 'field',
+                label: 'Field',
+                type: 'string',
+                input: 'text',
+                operators: ['field_equal', 'field_not_equal', 'field_greater', 'field_less', 'field_between']
+            }, {
+                id: '__current_date__',
+                label: 'Current Date',
+                type: 'date',
+                validation: {
+                    format: 'DD-MM-YYYY' // moment.js format
+                },
+                plugin: 'flatpickr',
+                plugin_config: {
+                    enableTime: false,
+                    dateFormat: 'd-m-Y', // flatpickr format
+                    time_24hr: true,
+                    weekNumbers: true,
+                    locale: {
+                        firstDayOfWeek: 1
+                    },
+                },
+                operators: ['equal', 'not_equal', 'less', 'less_or_equal', 'greater', 'greater_or_equal', 'between_dates']
+            }, {
+                id: '__current_date_time__',
+                label: 'Current Datetime',
+                type: 'date',
+                validation: {
+                    format: 'DD-MM-YYYY HH:mm' // moment.js format
+                },
+                plugin: 'flatpickr',
+                plugin_config: {
+                    enableTime: true,
+                    dateFormat: 'd-m-Y H:i', // flatpickr format
+                    time_24hr: true,
+                    weekNumbers: true,
+                    locale: {
+                        firstDayOfWeek: 1
+                    },
+                },
+                operators: ['less', 'less_or_equal', 'greater', 'greater_or_equal', 'between_dates']
+            }, {
+                id: '__current_time__',
+                label: 'Current Time',
+                type: 'date',
+                validation: {
+                    format: 'HH:mm' // moment.js format
+                },
+                plugin: 'flatpickr',
+                plugin_config: {
+                    enableTime: true,
+                    dateFormat: 'H:i', // flatpickr format
+                    time_24hr: true,
+                    noCalendar: true,
+                },
+                operators: ['less', 'less_or_equal', 'greater', 'greater_or_equal', 'between_dates']
+            }
+        ],
+        // rules: rules_basic
+    };
+
+    var rules = null;
+
+    try {
+        if ($(jquerBuilderJsonInput).val()) {
+            rules = JSON.parse($(jquerBuilderJsonInput).val());
+        }
+    } catch (error) {
+        console.log('Rules cannto be parsed');
+    }
+
+    if (rules) {
+        // load the rules if they exist
+        queryStructure['rules'] = rules;
+    }
+
+    $('.condition_builder').queryBuilder(queryStructure);
+
+    //When rules changed :
+    // $('.condition_builder').each(function () {
+    //     $(this).on('rulesChanged.queryBuilder', function (evt) {
+    //         var rules = $(this).queryBuilder('getRules');
+    //         if (rules) {
+    //             var jsonLogic = rulesToJsonLogic(rules);
+    //             monacoEditor.getModel().setValue(JSON.stringify(jsonLogic, null, 3));
+    //         }
+    //     });
+    // })
+}
+
+// show the data config builder
+// on click the modal is loaded and show the builder
+// on change it updates the monaco editor and the monaco editor updates the input fields
+function showConditionBuilder(json, monacoEditor, jquerBuilderJsonInput) {
+    var editor;
+    $('.conditionBuilderBtn').each(function () {
+        $(this).click(() => {
+            $(".condition_builder_modal_holder").modal({
+                backdrop: false
+            });
+            if (editor) {
+                // set the latest value if the user changed the JSON manually                
+                // editor.setValue(getJson(json));
+            }
+            $('.condition_builder_modal_holder').on('hidden.bs.modal', function (e) {
+                // on modal close set the value to the Monaco editor
+            })
+            $('.saveConditionBuilder').each(function () {
+                $(this).attr('data-dismiss', 'modal');
+                $(this).click(function () {
+                    var rules = $('.condition_builder').queryBuilder('getRules');
+                    $(jquerBuilderJsonInput).val(JSON.stringify(rules));
+                    monacoEditor.getModel().setValue(JSON.stringify(rulesToJsonLogic(rules), null, 3));
+                })
+            });
+        });
+    });
+
+    // get groups and prepare the consition builder
+    getGroups(jquerBuilderJsonInput, monacoEditor);
+
+    // var tree = JSON.parse('{"id":"ab989b98-0123-4456-b89a-b17d339bc6e3","type":"group","children1":{"baabba8b-4567-489a-bcde-f17d33a00d1a":{"type":"rule","properties":{"valueError":[null],"field":"user.firstName","operator":"equal","value":["s"],"valueSrc":["value"],"valueType":["text"]}}}}');
+    // jsonLogicFormat(tree,{});
+
+}
+
+// ********************************************* CONDITION BUILDER *****************************************
+
+// convert queryBuilder rules to JSON logic that we can use
+function rulesToJsonLogic(rules) {
+
+    rules = JSON.parse(JSON.stringify(rules).replace('"AND"', '"and"').replace('"OR"', '"or"'));
+
+    jsonLogic = {};
+
+    if (!rules) {
+        return jsonLogic;
+    }
+
+    jsonLogic[rules.condition] = [];
+    rules.rules.forEach(rule => {
+
+        var valuePrefix = rule.field == 'user_group' ? '$' : ''; // if the filed is user group we add the $ prefix
+
+        // get date and time formats if fields need it
+        var flatpickrMomentFormat = 'DD-MM-YYYY HH:mm';
+        var momentFormat = 'YYYY-MM-DD HH:mm';
+        if (rule.field == '__current_date__') {
+            flatpickrMomentFormat = 'DD-MM-YYYY';
+            momentFormat = 'YYYY-MM-DD';
+        } else if (rule.field == '__current_time__') {
+            flatpickrMomentFormat = 'HH:mm';
+            momentFormat = 'HH:mm';
+        }
+
+        if (['in', 'not_in'].includes(rule.operator)) {
+            rule.value.forEach(val => {
+                var r = {};
+                r[jsonLogicOperators[rule.operator].op] = [true, valuePrefix + val]
+                jsonLogic[rules.condition].push(r);
+            });
+        } else if (rule.operator == 'in_one_of') {
+            // add additional OR for one of these groups
+            jsonLogic[rules.condition] = {
+                "or": []
+            }
+            rule.value.forEach(val => {
+                var r = {};
+                r[jsonLogicOperators[rule.operator].op] = [true, valuePrefix + val]
+                jsonLogic[rules.condition]['or'].push(r);
+            });
+        }
+        else if (['field_equal', 'field_not_equal', 'field_greater', 'field_less'].includes(rule.operator)) {
+            var r = {};
+            r[jsonLogicOperators[rule.operator].op] = [rule.value[0], rule.value[1]]
+            jsonLogic[rules.condition].push(r);
+        } else if (rule.operator == 'field_between') {
+            // add additional AND for one of these groups
+            jsonLogic[rules.condition] = {
+                "and": []
+            }
+            jsonLogic[rules.condition]['and'].push({
+                ">=": [rule.value[0], rule.value[1]]
+            });
+            jsonLogic[rules.condition]['and'].push({
+                "<=": [rule.value[0], rule.value[2]]
+            });
+        } else if (rule.operator == 'between_dates') {
+            // add additional AND for one of these groups            
+            jsonLogic[rules.condition] = {
+                "and": []
+            }
+            jsonLogic[rules.condition]['and'].push({
+                ">=": [rule.field, new moment(rule.value[0], flatpickrMomentFormat).format(momentFormat)]
+            });
+            jsonLogic[rules.condition]['and'].push({
+                "<=": [rule.field, new moment(rule.value[1], flatpickrMomentFormat).format(momentFormat)]
+            });
+        } else if (['equal', 'not_equal', 'less', 'less_or_equal', 'greater', 'greater_or_equal'].includes(rule.operator)) {
+            var r = {};
+            var val = rule.value;
+            if (['__current_date__', '__current_date_time__', '__current_time__'].includes(rule.field)) {
+                // if needed convert date
+                val = new moment(rule.value, flatpickrMomentFormat).format(momentFormat);
+            }
+            r[jsonLogicOperators[rule.operator].op] = [rule.field, val]
+            jsonLogic[rules.condition].push(r);
+        }
+    });
+
+    return jsonLogic;
 }