+++ server/component/style/textarea/tpl_condition_builder.php
@@ -3,5 +3,5 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 ?>
-<div class="json form-control p-0"></div>
-<?php $this->output_builder() ?>
+<button type="button" class="w-100 mt-1 btn btn-primary conditionBuilderBtn btn-sm">Condition Builder</button>
+<div id="builder"></div>